+++ scraper/test-new-sources.ts
@@ -22,6 +22,7 @@ import { scrapeHiringPlatform } from './engines/hiringplatform';
 import { scrapeADP } from './engines/adp';
 import { scrapeApplyToEducation } from './engines/applytoeducation';
 import { scrapeWorkable } from './engines/workable';
+import { scrapeAvanti } from './engines/avanti';
 import { scrapeBrassRing, scrapeEdmontonPhenom, scrapeHaltonHills, scrapeNipissing, scrapeNorthernCollege, scrapePickering, scrapeStClairCollege, scrapeStLawrenceCollege, scrapeVaughanPL } from './engines/custom';
 
 const REQUIRED_SUCCESSFUL_RUNS = 3;
@@ -122,6 +123,8 @@ const SOURCES = {
     scrapeApplyToEducation(db, context, 'https://www.tcdsb.org/page/jobs', 'Toronto Catholic District School Board'),
   'Norfolk County': (db: Client, context: BrowserContext) =>
     scrapeWorkable(db, context, 'norfolk-county', 'Norfolk County'),
+  'Bruce County': (db: Client, context: BrowserContext) =>
+    scrapeAvanti(db, context, 'https://brucecounty.myavanti.ca/careers', 'Bruce County'),
   'Government of Alberta': (db: Client, context: BrowserContext) =>
     scrapeJobs2Web(db, context, 'https://jobpostings.alberta.ca/go/All-Jobs-GoA/2617217/', 'Government of Alberta'),
   'Ontario Energy Board': (db: Client, context: BrowserContext) =>